aiopsl : Support for CSCN/BPSCN task creation
WARNING : Requires MC version 10.9.0 or newer

- Configured the last 32 EPID table entries for the CSCN, BPSCN tasks
- Added API related to configuration of CSCN, BPSCN tasks execution
environment.
- Extended some existing API to allow configuration of BP depletion,
CG congestion notifications.
- Add API to retrieve the FQID. API is needed for the FDMA XOFF/XON
command.

+++ src/drivers/include/fsl_dpbp.h
@@ -395,7 +395,7 @@ struct dpbp_notification_cfg {
 	uint32_t	surplus_exit;
 	uint64_t	message_iova;
 	uint64_t	message_ctx;
-	uint16_t	options;
+	uint32_t	options;
 };
 
 /**
